requirements.txt
fastapi
uvicorn
tensorflow-cpu
pillow
numpy
requests